\name{caMachineLearning-package}
\Rdversion{1.1}
\alias{caMachineLearning-package}

\docType{package}

\title{
RWebServices interface to the XXX and YYY packages.
}

\description{
Provides an RWebServices interface to a part of the functionality provided
by the R XXX and YYY packages.
}

\details{
See 'examples' for package details.
}

\author{
Tom Milac <tmilac@fhcrc.org>
}

%\references{
%Literature or other references for background information
%}
\examples{
packageDescription('caMachineLearning')
}
\keyword{ package }

%\seealso{
%Optional links to other man pages, e.g.
%\code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}

%\examples{
%simple examples of the most important functions
%}
