\name{MachineLearningParameters-class}
\docType{class}
\alias{MachineLearningParameters-class}

\title{Class "MachineLearningParameters"}
\description{
  MachineLearningParameters is the base class for all machine learning parameter classes.
}
\section{Objects from the Class}{

  Objects can be created by calls to the constructor
  \code{\link{MachineLearningParameters}}.

}

%\section{Slots}{
%  \describe{
%  }
%}

\section{Methods}{

  See \code{\link{MachineLearningParameters}}

}

\author{Tom Milac \url{tmilac@fhcrc.org}}

\seealso{

  \code{\link{MachineLearningParameters}} for object construction, accessors

}

\examples{
showClass("MachineLearningParameters")
obj <- MachineLearningParameters()
}
\keyword{classes}
