\name{MachineLearningParameters}
\alias{MachineLearningParameters}

\title{MachineLearningParameters constructor and accessors}

\description{
  Functions for creating and manipulating MachineLearningParameters-class objects.
}

\usage{
MachineLearningParameters()
}

%\arguments{
%#}

\value{

  \item{MachineLearningParameters}{Object of class
    \code{\linkS4class{MachineLearningParameters}}.}

}

\author{Tom Milac \url{tmilac@fhcrc.org}}

\seealso{\code{\linkS4class{MachineLearningParameters}-class}}

\examples{
obj <- MachineLearningParameters()
}

\keyword{manip}
